backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.123.9",
    "pymongo[srv]>=4.15.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
